feat(k8s/data): add new data types based on SIG plans (#1)
In the 11/2-11/3/2016 sig-service-catalog discussions in Boulder at the
Deis offices, the group came to consensus on what Kubernetes resources
to use (the nouns) and how they’ll be used (the API). Both concepts are
captured briefly in
https://github.com/kubernetes-incubator/service-catalog/pull/31.

This PR adds the nouns to the steward framework codebase. Future
patches will begin incorporating them into the state machine to address
the API changes